+++ emu/nds/rast/gl/vector.go
@@ -5,10 +5,10 @@ import (
 )
 
 type Vector struct {
-	X, Y, Z float64
+	X, Y, Z float32
 }
 
-func (a Vector) Dot(b Vector) float64 {
+func (a Vector) Dot(b Vector) float32 {
 	return a.X*b.X + a.Y*b.Y + a.Z*b.Z
 }
 
@@ -20,28 +20,40 @@ func (a Vector) Sub(b Vector) Vector {
 	return Vector{a.X - b.X, a.Y - b.Y, a.Z - b.Z}
 }
 
-func (a Vector) MulScalar(b float64) Vector {
+func (a Vector) MulScalar(b float32) Vector {
 	return Vector{a.X * b, a.Y * b, a.Z * b}
 }
 
 func (a Vector) Min(b Vector) Vector {
-	return Vector{math.Min(a.X, b.X), math.Min(a.Y, b.Y), math.Min(a.Z, b.Z)}
+	return Vector{
+        float32(math.Min(float64(a.X), float64(b.X))),
+        float32(math.Min(float64(a.Y), float64(b.Y))),
+        float32(math.Min(float64(a.Z), float64(b.Z)))}
 }
 
 func (a Vector) Max(b Vector) Vector {
-	return Vector{math.Max(a.X, b.X), math.Max(a.Y, b.Y), math.Max(a.Z, b.Z)}
+	return Vector{
+        float32(math.Max(float64(a.X), float64(b.X))),
+        float32(math.Max(float64(a.Y), float64(b.Y))),
+        float32(math.Max(float64(a.Z), float64(b.Z)))}
 }
 
 func (a Vector) Floor() Vector {
-	return Vector{math.Floor(a.X), math.Floor(a.Y), math.Floor(a.Z)}
+	return Vector{
+        float32(math.Floor(float64(a.X))),
+        float32(math.Floor(float64(a.Y))),
+        float32(math.Floor(float64(a.Z)))}
 }
 
 func (a Vector) Ceil() Vector {
-	return Vector{math.Ceil(a.X), math.Ceil(a.Y), math.Ceil(a.Z)}
+	return Vector{
+        float32(math.Ceil(float64(a.X))),
+        float32(math.Ceil(float64(a.Y))),
+        float32(math.Ceil(float64(a.Z)))}
 }
 
 type VectorW struct {
-	X, Y, Z, W float64
+	X, Y, Z, W float32
 }
 
 func (a VectorW) Vector() Vector {
@@ -53,11 +65,11 @@ func (a VectorW) Outside() bool {
 	return x < -w || x > w || y < -w || y > w || z < -w || z > w
 }
 
-func (a VectorW) Dot(b VectorW) float64 {
+func (a VectorW) Dot(b VectorW) float32 {
 	return a.X*b.X + a.Y*b.Y + a.Z*b.Z + a.W*b.W
 }
 
-func (a VectorW) Dot3(b VectorW) float64 {
+func (a VectorW) Dot3(b VectorW) float32 {
 	return a.X*b.X + a.Y*b.Y + a.Z*b.Z
 }
 
@@ -69,10 +81,10 @@ func (a VectorW) Sub(b VectorW) VectorW {
 	return VectorW{a.X - b.X, a.Y - b.Y, a.Z - b.Z, a.W - b.W}
 }
 
-func (a VectorW) MulScalar(b float64) VectorW {
+func (a VectorW) MulScalar(b float32) VectorW {
 	return VectorW{a.X * b, a.Y * b, a.Z * b, a.W * b}
 }
 
-func (a VectorW) DivScalar(b float64) VectorW {
+func (a VectorW) DivScalar(b float32) VectorW {
 	return VectorW{a.X / b, a.Y / b, a.Z / b, a.W / b}
 }
